Fixed unsigned types and error msg in sys_spec.c
Fixed an issue with sys_spec.c as the unsigned keyword was not being
added to the typedefs being output.

Also fixed an error message stating that an error occured when the
system is not a 32-bit system when the error only occurs on systems that
are not 32 OR 64 bit.